requirements.txt
flask
flask_sqlalchemy
python-telegram-bot==13.15
pandas
openpyxl
edge-tts
